requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
# 新增diff相关依赖
unified-diff>=0.3.0
patch>=1.16.0 